Pass global config object to DataManager.
Purpose of this commit:
The DataManager needs to have the database created and its schema in place before the queue consumer thread can be started, and the queue consumer thread needs to be started before the plugins can be started. As a result, it's necessary to read the complete configuration file before starting the plugins. Additionally, the DataManager needs a path to the database file.

* The GlobalConfig class now parses and stores the configuration file data (which includes the config data for each plugin) into a dictionary of dictionaries.
* The GlobalConfig class instance is now passed to the DataManager so that it can access each plugin's config, which includes the plugin's table name and column definitions.
* After the DataManager is instantiated, the Framework iterates the config dictionary to discover, load and start enabled plugins.
* Unit tests have been updated subsequent to these changes.

The GlobalConfig object now exposes three key methods:
* get_db_path() - returns the (fully qualified, absolute path) to the database file
* get_ports() - returns a list of ports with enabled plugins associated with them
* get_plugin_config(port) - returns the config dictionary for the specified port

Thus it is possible to iterate the enabled plugin configs, by iterating the list returned by get_ports() and then accessing the individual plugin configs with get_plugin_config(port).