Lots of ugly code added, it can create a database and tables from the configs. Not perfect but it does seem to work.
I can import the framework, create a framework object,
start the framework object which will in turn read the config
create the database, create the tables and the fields, it
will then spin up the plugins and I can also call the insert
method in the framework and have the data manager insert
data into the database.

Fully qualified some names to make things work and added a small
method for getting the folder name for the DB in the framework.

I will give a demo tomorrow if anyone wants. Only pushing to
my remote and not doing a pull request until we all get a closer
look at it.