Flattened directory structure.
This may look like a radical change, but it's not; it's simply a flatter structure than what we had before. I've left the "dashboard" and "webserver" directories intact. I ran the webserver and it seems happy. Save for the unit tests, it's no longer necessary to set the python path when starting the framework - the recce7.sh script is there, but invoking the framework like a python program with 'python3 -m framework.frmwork' works as expected too. We will be using the recce7.sh script for setting RECCE7_*** environment variables and other magic. Right now it just invokes the framework.

I've moved all unit tests to the tests/ directory. A structure similar to the recce7/ directory is there, only with test files in it instead. We can store mock configurations or whatever there as well.

This should make it easier for everyone to use JetBrains IDEs (either PyCharm or IntelliJ IDEA + Python plugin). You no longer need to mark any directories as sources; just run frmwork.py to start the framework. It's also easier to specify unit tests, just set the project root to the main recce7 directory and tell it to run anything matching the pattern *_test.py .

Lastly, I've moved globalconfig.py to its own directory, in common/. Code that's shared between the framework and the report server can go there now.